+++ pyplotter/sources/widgetCSV.py
@@ -89,7 +89,7 @@ class WidgetCSV(QtWidgets.QWidget):
     signalLabelSnapshotEnabled       = QtCore.pyqtSignal(bool)
     signalAddSnapshot = QtCore.pyqtSignal(dict)
 
-    signalUpdateProgressBar = QtCore.pyqtSignal(QtWidgets.QProgressBar, int, str)
+    signalUpdateProgressBar = QtCore.pyqtSignal(QtWidgets.QProgressBar, float, str)
     signalRemoveProgressBar = QtCore.pyqtSignal(QtWidgets.QProgressBar)
     signalFillTableWidgetParameter = QtCore.pyqtSignal(int, list, dict, dict, str, str, str, str, bool)
     signalLoadedDataFull = QtCore.pyqtSignal(int, str, str, str, str, str, QtWidgets.QCheckBox, QtWidgets.QProgressBar, tuple, str, str, str, str, str, str, bool)
@@ -229,7 +229,7 @@ class WidgetCSV(QtWidgets.QWidget):
                        cb: QtWidgets.QCheckBox,
                        progressBar: QtWidgets.QProgressBar) -> None:
 
-        self.signalUpdateProgressBar.emit(progressBar, 100, 'Downloading data: 100%')
+        self.signalUpdateProgressBar.emit(progressBar, 100., 'Downloading data: 100%')
 
         for paramDependent in self.paramDependentList:
             if paramDependent['name'] == dependentParamName:
